Updated the recombining algorithm to pixel-wise weighting.
The images from each biased CNN are combined more intelligently, through
pixel-wise weighting. For any given pixel in the 3 images, we compute
the intensity (sum of channels) of that pixel. The pixel is then
weighted relative to the pixel intensities from the other two biases.
These weights are used to sum each pixel of the image, generating the
output image.