Updated the testing script to use a fourth biased model (blue-green).
The testing script now assumes 4 biased models: red, blue, green, and a
combined blue and green biased model. To deal with this, the saturations
of the models are also weighted to account for the fact that there are
more blue-green biased models.

Also, print messages were cleaned up, and a missing directory creation
for the output directory was added.